requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.8.0
click==8.1.8
colorama==0.4.6
db==0.1.1
fastapi==0.115.8
h11==0.14.0
idna==3.10
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
